Don't bind lockd to specific IPs.
Ticket:	#14957

Like many other services NFS has a provision to bind
to specific IPs.  This is done by passing -h IP flags
to the daemons in the NFS stack.  rpcbind, nfsd, mountd,
statd, and lockd all take these flags.  You can pass
multiple -h flags to the daemons and they will listen
to every IP that you pass them as well as 127.0.0.1 and ::1

Except for lockd.  Lockd has a bug that apparently has existed
in FreeBSD for a Very Long Time where it only listens to
the last IP that you pass it.  In most cases where you only
have one bind IP this isn't a problem, lockd listens to that
one IP just fine.  If you don't specify a bind IP lockd
listens on 0.0.0.0 which also works fine.  If you specify
more than one IP lockd only listens to the last IP you
specify which can cause hilarity. (where hilarity is defined
as lockd not listening where you tell it to)

This hackaround has the NFS stack listen on specific IPs
as it always has, but has lockd listen on 0.0.0.0.  It's better
than not having lockd work at all but worse than having it listen
in the right places.